dom/tests/mochitest/bugs/test_bug414291.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=414291
-->
<head>
  <title>Test for Bug 414291</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=414291">Mozilla Bug 414291</a>
<script class="testbody" type="text/javascript">

/** Test for Bug 414291 **/
var result1 = 0;
var result2 = 0;

window.open("file1_bug414291.html", "w1");
is(result1, 0, "window should not be opened either as modal or loaded synchronously.");

window.open("file2_bug414291.html", "w2", "modal=yes");
is(result2, 0, "window should not be opened either as modal or data loaded synchronously.");

</script>
</pre>
</body>
</html>
